+++ src/main/java/core/model/FolderUtil.java
@@ -1,4 +1,4 @@
-package core.contentManager;
+package core.model;
 
 public class FolderUtil {
