+++ src/main/java/swing/objects/core/PanelType.java
@@ -1,6 +1,6 @@
 package swing.objects.core;
 
-import swing.objects.objects.WrapLayout;
+import swing.util.WrapLayout;
 
 import java.awt.*;
